Emitter is working real good, however we need to improve it and there's a problem with some gpu buffer apparantly, and that buffer(s) is/are filling up entirely and all my objects go black (material data isnt processed at all

+++ particle_system_simulator/src/MeshConstruction/Shapes.cpp
@@ -221,7 +221,7 @@ Mesh createSphere(uint32_t approximateVertexCount, bool smooth)
 	{
 		a = 1.0f;
 		b = 2.0f;
-		c = 1 - approximateVertexCount;
+		c = 1 - static_cast<int>(approximateVertexCount);
 	}
 	else
 	{
